+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-16T08:43:25.167Z",
+    "2026-02-16T11:16:38.883Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {